Fix strict mode
