+++ src/Explorer/Menus/NotificationConsole/ConsoleData.tsx
@@ -1,0 +1,16 @@
+/**
+ * Interface for the data/content that will be recorded
+ */
+
+export interface ConsoleData {
+  type: ConsoleDataType;
+  date: string;
+  message: string;
+  id?: string;
+}
+
+export enum ConsoleDataType {
+  Info = 0,
+  Error = 1,
+  InProgress = 2,
+}
